Function micro_rpc_build::compile

pub fn compile(
    protos: &[impl AsRef<Path>],
    includes: &[impl AsRef<Path>],
    options: CompileOptions
)

Compile Rust server code from the services in the provided protobuf file.

Each method in a service definition must have exactly one comment line of the form // method_id: 42, which is used to generate a stable identifier for that method, which is part of the serialization protocol used for invocation.

For a service called TestName, compile generates the following objects:

a struct named TestNameServer, which implements the micro_rpc::InvocationHandler trait, dispatching each request to the appropriate method on the underlying service implementation.

a trait named TestName, with a method for each method defined in the protobuf service, and an additional default method named serve which returns an instance of TestNameServer; the developer of a service would usually define a concrete struct and manually implement this trait for it,

a struct named TestNameClient, exposing a method for each method defined in the protobuf service. This may be used to directly invoke the underlying handler in order to indirectly invoke methods on the corresponding Server object on the other side of the handler.

a struct named TestNameAsyncClient, similar to TestNameClient but with async support.